+++ packages/dso-toolkit/src/components/footer/footer.models.ts
@@ -1,0 +1,3 @@
+export interface Footer<TemplateFnReturnType> {
+  children: TemplateFnReturnType;
+}