+++ storybook/src/example-pages/snippets/table/table-data-table.ts
@@ -1,4 +1,0 @@
-import { examplePageFactory } from "../../../example-page-factory";
-import { content } from "./table-snippet.content";
-
-examplePageFactory("Snippets/Table", "Data Table", ({ tableTemplate }, templates) => tableTemplate(content(templates)));