+++ storybook/src/example-pages/toepassingen/regels-op-de-kaart/landingspagina.ts
@@ -7,6 +7,7 @@ import { header } from "../../partials/header.content";
 import { accordionSections } from "./landingspagina.content";
 
 examplePageFactory(
+  "Voorbeeldpagina's",
   "Toepassingen/Regels op de kaart",
   "Landingspagina",
   ({ accordionTemplate, anchorTemplate, highlightBoxTemplate, richContentTemplate }, templates) => html`